+++ modules/core/src/persistence_yml.cpp
@@ -573,7 +573,7 @@ public:
             else
             {
             force_int:
-                int ival = (int)strtol( ptr, &endptr, 0 );
+                int64_t ival = strtol( ptr, &endptr, 0 );
                 node.setValue(FileNode::INT, &ival);
             }
